Fix bionic regression - midolman uses 100% CPU
The WakerUpper uses a busy loop with a sleep to continually check
the status of "parked" threads.  The sleep used to be a nanosecond
sleep of 1 nanosecond ... with a comment that said this was 50
microsecond "on recent kernels".  With Bionic, this is no longer
true, and the WakerUpper thread runs 99% of the time.

This fix changes the sleep to 50,000 nanoseconds, restoring the
old Xenial performance characteristics.  (Roughly 30% CPU usage.)

Change-Id: I34a01115c4b1998598f5d1212e078c93ff039da7